+++ pkg/gatewayserver/io/mock/frontend.go
@@ -43,7 +43,7 @@ func ConnectFrontend(ctx context.Context, ids *ttnpb.GatewayIdentifiers, server 
 		Down:   make(chan *ttnpb.DownlinkMessage, 1),
 	}
 
-	ctx = io.NewContextWithGatewayRemoteAddress(ctx, &ttnpb.GatewayRemoteAddress{IP: "127.0.0.1"})
+	ctx = io.NewContextWithGatewayRemoteAddress(ctx, &ttnpb.GatewayRemoteAddress{Ip: "127.0.0.1"})
 
 	conn, err := server.Connect(ctx, f, ids)
 	if err != nil {