+++ pkg/gatewayserver/io/udp/udp.go
@@ -220,7 +220,7 @@ func (s *srv) connect(ctx context.Context, eui types.EUI64, addr *net.UDPAddr) (
 		})
 
 		ctx = io.NewContextWithGatewayRemoteAddress(ctx, &ttnpb.GatewayRemoteAddress{
-			IP: addr.IP.String(),
+			Ip: addr.IP.String(),
 		})
 		conn, err = s.server.Connect(ctx, s, ids)
 		if err != nil {